Backend API: Improve contract and transaction handling
- Updated contract storage logic to avoid unnecessary encoding.
- Improved transaction retrieval by handling hex formatting properly.
- Fixed address lookup to create missing entries dynamically.
- Enhanced transaction counting logic for performance.
- Refactored routes to use more consistent variable names.
- Updated models for contract and address storage.

+++ backendAPI/models/address.go
@@ -5,6 +5,6 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type Address struct {
 	ObjectId primitive.ObjectID `bson:"_id"`
 	ID       string             `json:"id"` // Changed from []byte to string
-	Balance  float32            `json:"balance"`
+	Balance  float64            `json:"balance"`
 	Nonce    uint64             `json:"nonce"`
 }
